backend/requirements.txt
# ReconXploit Backend Requirements

# Web Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Database
psycopg2-binary==2.9.9
sqlalchemy==2.0.23
alembic==1.12.1

# Task Queue
celery==5.3.4
redis==5.0.1
flower==2.0.1

# HTTP Client
httpx==0.25.1
requests==2.31.0
aiohttp==3.9.0

# Data Processing
pydantic==2.5.0
pydantic-settings==2.1.0
python-dotenv==1.0.0

# Parsing & Analysis
beautifulsoup4==4.12.2
lxml==4.9.3
xmltodict==0.13.0

# Security Tools Integration
python-nmap==0.7.1
dnspython==2.4.2

# Utilities
pyyaml==6.0.1
python-dateutil==2.8.2
pytz==2023.3

# Notifications
python-telegram-bot==20.7
discord-webhook==1.3.0

# File handling
openpyxl==3.1.2
reportlab==4.0.7
markdown==3.5.1

# CLI
click==8.1.7
rich==13.7.0
colorama==0.4.6

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
httpx==0.25.1

# Code Quality
black==23.11.0
flake8==6.1.0
mypy==1.7.1

# Monitoring
prometheus-client==0.19.0

# Async
asyncio==3.4.3
aiofiles==23.2.1
